# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: keel API
  version: 1.0.0
  description: |
    Contract-first API (anayasa K3, ADR-006, ADR-024). The backend implements this file; the mobile client's types are
    generated from it, never hand-written. Rules the contract keeps, each checked by backend ContractTests:
    - The Decision is the engine's Decision field for field (U3). Actions are a tagged union on `type`.
    - Every calorie estimate is a range, `KcalRange` / `KcalBalance` (U5). A target and a decision's step may be one
      number (U5); a device's own reading is passed through as given.
    - The engine's fat estimate never appears, in any form (U4). Error messages never carry health data (V3).
    Conventions: everything but health and sign-in is under /v1 and needs a session. Records the phone creates
    offline carry a `clientId` (a UUID the phone makes): sending one twice stores it once and answers 200 with the
    stored record instead of 201. Days are `date` (the user's local calendar day); moments are `date-time` with offset.
    Lists page backwards: `limit` and `before` (the cursor of the last item seen).
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - session: []
...
paths:
...
  /v1/exercises:
    get:
      tags: [training]
      operationId: listExercises
      summary: The exercise catalog
      responses:
        "200":
          description: Every exercise
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Exercise"
        default:
          $ref: "#/components/responses/Error"
...
  /v1/workouts:
    get:
      tags: [training]
      operationId: listWorkouts
      parameters:
        - $ref: "#/components/parameters/From"
        - $ref: "#/components/parameters/To"
      responses:
        "200":
          description: Workouts in the range, oldest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Workout"
        default:
          $ref: "#/components/responses/Error"
    post:
      tags: [training]
      operationId: startWorkout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewWorkout"
      responses:
        "201":
          $ref: "#/components/responses/WorkoutCreated"
        "200":
          $ref: "#/components/responses/WorkoutCreated"
        default:
          $ref: "#/components/responses/Error"
  /v1/workouts/{id}:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      tags: [training]
      operationId: getWorkout
      responses:
        "200":
          description: The workout with its sets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workout"
        default:
          $ref: "#/components/responses/Error"
  /v1/workouts/{id}/finish:
    parameters:
      - $ref: "#/components/parameters/Id"
    post:
      tags: [training]
      operationId: finishWorkout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkoutFinish"
      responses:
        "200":
          description: The finished workout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workout"
        default:
          $ref: "#/components/responses/Error"
  /v1/workouts/{id}/sets:
    parameters:
      - $ref: "#/components/parameters/Id"
    post:
      tags: [training]
      operationId: logSet
      description: A clientId already used for a set of another workout is 409 CONFLICT, not a replay.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewSet"
      responses:
        "201":
          $ref: "#/components/responses/SetCreated"
        "200":
          $ref: "#/components/responses/SetCreated"
        default:
          $ref: "#/components/responses/Error"
  /v1/workouts/{id}/sets/{setId}:
    parameters:
      - $ref: "#/components/parameters/Id"
      - name: setId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    delete:
      tags: [training]
      operationId: deleteSet
      responses:
        "204":
          description: Deleted
        default:
          $ref: "#/components/responses/Error"

  # ── check-in (K-213) ───────────────────────────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    session:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: The access token of a Session (K-203).
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    From:
      name: from
      in: query
      required: true
      description: First day, inclusive (local day). A range is at most 400 days.
      schema:
        type: string
        format: date
    To:
      name: to
      in: query
      required: true
      description: Last day, inclusive
      schema:
        type: string
        format: date
...
  responses:
    Error:
      description: Any error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    WorkoutCreated:
      description: Started (201), or already started with this clientId (200)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Workout"
    SetCreated:
      description: Stored (201), or already stored with this clientId (200)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/LoggedSet"
  schemas:
...
    Error:
      type: object
      description: Shared error body. Messages never contain health data (anayasa V3).
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable machine-readable code, e.g. VALIDATION_FAILED, NOT_FOUND, CONSENT_REQUIRED.
        message:
          type: string
          description: English, safe to show to the user.
...
    ClientId:
      type: string
      format: uuid
      description: Made by the phone when the record is created; the same record sent again is stored once.

    # ── auth ─────────────────────────────────────────────────────────────────────────────────────────────────
...
    Exercise:
      type: object
      required: [id, nameKey, kind, muscles, alternatives]
      properties:
        id:
          type: string
        nameKey:
          type: string
          description: Copy key of the name (en.json); aliases are for search.
        aliases:
          type: array
          items:
            type: string
        kind:
          type: string
          enum: [COMPOUND, ISOLATION]
        muscles:
          type: array
          items:
            type: string
        alternatives:
          type: array
          description: Exercise ids that can take its place.
          items:
            type: string
        load:
          type: string
          enum: [EXTERNAL, BODYWEIGHT, BODYWEIGHT_PLUS_EXTERNAL]
        unilateral:
          type: boolean
        setupFields:
          type: array
          description: What the user sets on the machine and keeps on the phone (seat, pad, grip; ADR-017).
          items:
            type: string
        clips:
          $ref: "#/components/schemas/ExerciseClips"
    ExerciseClips:
      type: object
      description: The two demonstration clips (ADR-017) — the first rep and the last rep near the target RIR.
      required: [firstRep, lastRep]
      properties:
        firstRep:
          type: string
        lastRep:
          type: string
...
    NewWorkout:
      type: object
      required: [clientId, startedAt]
      properties:
        clientId:
          $ref: "#/components/schemas/ClientId"
        startedAt:
          type: string
          format: date-time
        programDayId:
          type: string
          format: uuid
    WorkoutFinish:
      type: object
      required: [endedAt]
      properties:
        endedAt:
          type: string
          format: date-time
    Workout:
      type: object
      required: [id, clientId, startedAt, sets]
      properties:
        id:
          type: string
          format: uuid
        clientId:
          $ref: "#/components/schemas/ClientId"
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
        programDayId:
          type: string
          format: uuid
        sets:
          type: array
          items:
            $ref: "#/components/schemas/LoggedSet"
    SetType:
      type: string
      description: Only WORKING sets count toward effort and estimated 1RM (K-218).
      enum: [WARM_UP, WORKING, DROP, FAILURE]
    Side:
      type: string
      enum: [BOTH, LEFT, RIGHT]
    NewSet:
      type: object
      description: |
        One set. exerciseId may differ from the plan (a swap, e.g. pull-ups for lat pulldown). A bodyweight exercise
        logs the added load in loadKg (0 without); the engine adds the bodyweight.
      required: [clientId, exerciseId, setType, reps, loadKg]
      properties:
        clientId:
          $ref: "#/components/schemas/ClientId"
        exerciseId:
          type: string
        setType:
          $ref: "#/components/schemas/SetType"
        loadKg:
          type: number
          minimum: 0
          maximum: 1000
          description: At most 2 decimals.
        reps:
          type: integer
          minimum: 0
        rir:
          type: integer
          minimum: 0
          description: Reps left in the tank; absent when not given.
        side:
          $ref: "#/components/schemas/Side"
    LoggedSet:
      allOf:
        - $ref: "#/components/schemas/NewSet"
        - type: object
          required: [id]
          properties:
            id:
              type: string
              format: uuid

    # ── check-in ─────────────────────────────────────────────────────────────────────────────────────────────
...